lib/stash-python/pyproject.toml
[project]
name = "stash-client"
version = "0.1.0"
description = "Python client library for Stash KV service"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Umputun" }]
keywords = ["stash", "kv", "key-value", "configuration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "urllib3>=2.0",
    "cryptography>=41.0",
    "argon2-cffi>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "responses>=0.25",
]

[project.urls]
Homepage = "https://github.com/umputun/stash"
Documentation = "https://github.com/umputun/stash/tree/master/lib/stash-python"
Repository = "https://github.com/umputun/stash"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stash"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/stash"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
